2011-05-19 20:28 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * src/rtl/arc4.c     ! silence warning on bsd [Tamas Tevesz]

+++ harbour/src/rtl/arc4.c
@@ -276,7 +276,7 @@ static int arc4_seed_sysctl_bsd( void )
    }
 
    /* make sure that the buffer actually got set. */
-   for( i = any_set = 0; i < sizeof( buf ); ++i )
+   for( i = any_set = 0; i < ( int ) sizeof( buf ); ++i )
       any_set |= buf[ i ];
 
    if( ! any_set )
